Fix :display_as for texts including apostrophes (').
Since ' is used to identify the attribute's content, if it wasn't quoted
within it, it would end the attribute too soon causing both page validation
to go awry and any content following it to be dropped altogether.

Note that the ampersand (&) symbol has to be quoted or it could be
interpreted as the start of an XML entity.